requirements.txt
transformers>=4.40.0
torch>=2.2.0
sentencepiece>=0.2.0
customtkinter>=5.2.0
keyboard>=0.13.5
pyperclip>=1.8.2
pyautogui>=0.9.54
pystray>=0.19.5
Pillow>=10.0.0
